Adds new nimble ble code
