+++ CMakeLists.txt
@@ -3,6 +3,5 @@
 # The following five lines of boilerplate have to be in your project's
 # CMakeLists in this exact order for cmake to work correctly
 cmake_minimum_required(VERSION 3.16)
-set(EXTRA_COMPONENT_DIRS components/common/nimble_peripheral_utils)
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
 project(assistive-shoulder-firmware)
