--- components/common/nimble_central_utils/CMakeLists.txt
@@ -1,3 +1,0 @@
-idf_component_register(SRCS "misc.c" "peer.c"
-                       INCLUDE_DIRS "."
-                       PRIV_REQUIRES bt)
