--- components/common/nimble_peripheral_utils/CMakeLists.txt
@@ -1,3 +1,0 @@
-idf_component_register(SRCS "misc.c" "scli.c"
-                       INCLUDE_DIRS "."
-                       PRIV_REQUIRES bt console driver)
