--- components/common/nimble_peripheral_utils/esp_peripheral.h
@@ -1,31 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2021-2022 Espressif Systems (Shanghai) CO LTD
- *
- * SPDX-License-Identifier: Unlicense OR CC0-1.0
- */
-
-#ifndef H_ESP_PERIPHERAL_
-#define H_ESP_PERIPHERAL_
-
-#include <stdbool.h>
-#include "nimble/ble.h"
-#include "modlog/modlog.h"
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/* Console */
-int scli_init(void);
-int scli_receive_key(int *key);
-
-/** Misc. */
-void print_bytes(const uint8_t *bytes, int len);
-void print_addr(const void *addr);
-char *addr_str(const void *addr);
-void print_mbuf(const struct os_mbuf *om);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
