--- components/nimble-ble-prph/include/nimble-ble-prph.h
@@ -1,8 +1,0 @@
-//
-// Created by thedevmanek on 12/5/23.
-//
-
-#ifndef ASSISTIVE_SHOULDER_FIRMWARE_NIMBLE_BLE_CENT_H
-#define ASSISTIVE_SHOULDER_FIRMWARE_NIMBLE_BLE_CENT_H
-void init_ble();
-#endif //ASSISTIVE_SHOULDER_FIRMWARE_NIMBLE_BLE_CENT_H
