+++ components/nimble-prph/CMakeLists.txt
@@ -1,0 +1,3 @@
+idf_component_register(SRCS "nimble-prph.c" "misc.c"
+                    INCLUDE_DIRS "include"
+                    REQUIRES  nvs_flash bt)
