+++ components/nimble-prph/Kconfig.projbuild
@@ -48,33 +48,4 @@ menu "Example Configuration"
         help
             Use this option to enable/disable Security Manager Secure Connection 4.2 feature.
 
-    config EXAMPLE_EXTENDED_ADV
-        bool
-        depends on SOC_BLE_50_SUPPORTED
-        default y if SOC_ESP_NIMBLE_CONTROLLER
-        select BT_NIMBLE_EXT_ADV
-        prompt "Enable Extended Adv"
-        help
-            Use this option to enable extended advertising in the example.
-            If this option is disabled, ensure config BT_NIMBLE_EXT_ADV is
-            also disabled from Nimble stack menuconfig
-
-    config EXAMPLE_RANDOM_ADDR
-        bool
-        prompt "Advertise RANDOM Address"
-        help
-            Use this option to advertise a random address instead of public address
-
-    config EXAMPLE_ENCRYPTION
-        bool
-        prompt "Enable Link Encryption"
-        help
-            This adds Encrypted Read and Write permissions in the custom GATT server.
-
-    config EXAMPLE_RESOLVE_PEER_ADDR
-        bool
-        prompt "Enable resolving peer address"
-        help
-            Use this option to enable resolving peer's address.
-
 endmenu
