+++ main/CMakeLists.txt
@@ -1,3 +1,3 @@
 idf_component_register(SRCS "main.c"
                     INCLUDE_DIRS "."
-        REQUIRES nimble-ble-prph)
+        REQUIRES nimble-prph)
